mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "practitioner-seo"
version = "0.1.0"
description = "An opinionated, evidence-based SEO toolkit for Claude Desktop"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Practitioner SEO Contributors" },
]
keywords = ["seo", "mcp", "claude", "search-console", "serp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp>=1.25,<2",
    "httpx>=0.27,<1",
    "beautifulsoup4>=4.12,<5",
    "google-auth>=2.28,<3",
    "google-auth-oauthlib>=1.2,<2",
    "google-api-python-client>=2.120,<3",
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.scripts]
practitioner-seo = "practitioner_seo.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/practitioner_seo"]
